fix: 수익률 계산 오류 수정
totalProfit 을 int -> float 로 수정

+++ src/main/java/lotto/Application.java
@@ -127,7 +127,7 @@ class SingleGame {
 
 class Game{
     public String inputPrice;
-    public int totalProfit = 0;
+    public float totalProfit = 0;
     public int firstPrize = 0;
     public int secondPrize = 0;
     public int thirdPrize = 0;
